[NBKCoreKit] Some temporary allocation stuff.
Swift.withUnsafeTemporaryAllocation(of:capacity:) does not state that `count` equals `capacity` (although the current implementation makes it so). As a consequence, I have added a method with this guarantee. The fact that the `capacity` may be greater than the `count` is understandable, but not a headache I want to deal with. Having said that, I have made an exception for single element allocations using another method (for performance reasons). It returns a pointer, not a buffer, to circumvent this headache.