Cleanup (#114).
It should now be possible (in theory) to sieve every value in 0 through UInt.max on setup. This means that the sieve's stride may no longer fit in UInt, so I removed it. I also had to update some preconditions in Cache.init and Sieve/increment(). Would you ever sieve 0 through UInt.max on setup? Doubt it. But the restriction (in theory) was awkward.